Brings PR #353 (Qwen3.6-27B-NVFP4 on Strix Halo, gfx1151) up to date with
main so it stops being CONFLICTING — while it was, every pull_request
workflow silently skipped, so the branch has had no real CI since 07-22 and
has never run the windows/amd native-HIP row that landed after it (#358/#359).

Conflict resolutions, all preserving GB10 behaviour byte-for-byte:

- prefill_paged_compute.cuh: keep the branch's smem_K buffer-share (it removed
  smem_V, -8.25 KB LDS) AND main's (__bf16)(float) cast routing, which the
  Windows HIP SDK needs to disambiguate __hip_bfloat16's implicit operators.
  The fp16 arm two lines below had already auto-merged to exactly this form.
- mx_block_scale.cuh: take main's symlink into kernels/gb10 over the branch's
  regular-file copy. They are byte-identical, so the symlink is strictly better
  (one definition, not two).
- mtp_head.rs / speculative.rs: main split DraftProposer out to
  mtp_head/draft_proposer.rs and landed the same carry-drafter feature
  independently. main's copies are supersets (identical code, better comments),
  so the branch's inline versions go; nothing is lost.
- impl_a3.rs: keep BOTH lm-head arms. The Strix batch3 arm (K=3 verify) is now
  gated on cfg!(atlas_scale) — its kernel lives in the shared kernels/gb10 tree,
  so try_kernel resolves it on NVIDIA too, and without the gate GB10 would
  silently move num_tokens == 3 off main's batch4 arm.
- multi_seq/ffn.rs: drop the branch's n == 4 arm. main's (4..=8) arm is the
  general form of it and gates on can_forward_km at branch entry, which is
  precisely the half-done-fallthrough this one could hit.
- save_checkpoint.rs: take main's always-on is_prompt_tail predicate and keep
  the branch's is_tail as a SEPARATE opt-in (ATLAS_SSM_TAIL_CKPT, default off).
  The golden run never set that env var, so its is_tail was dead code there —
  nothing regresses either way.
- impl_a1.rs / mtp_carry.rs / types.rs / sequence.rs: main's, which route the
  policy through drafter_context (default-on + kill switch) instead of a raw
  env read. The golden serve set ATLAS_MTP_CARRY_DRAFTER=1 and
  --mtp-quantization bf16, so both resolve the same way.

Also fixed, all merge fallout:

- layer/tests.rs: both sides added midchunk_capture to the same literal.
- cuda_backend/copy2d.rs (new): the branch's cfg(atlas_scale) per-row fallback
  for copy_d2d_2d_async (SCALE and native-HIP link no cudart) pushed gpu_impl.rs
  to 508 LoC, over the 500 cap. Extracted the one vendor-forked method to a
  sibling module; gpu_impl.rs is now 469 and its trait body stays uniform.

Verified locally: cargo check + clippy --workspace --tests clean on both the
NVIDIA default and ATLAS_TARGET_HW=strix-hip configs, cargo fmt clean, and no
.rs over the 500-LoC cap.

+++ kernels/strix-hip/common/prefill_paged_compute.cuh
@@ -240,7 +240,7 @@ extern "C" __global__ void KERNEL_NAME(
                 v16bf a;
                 #pragma unroll
                 for (int i = 0; i < 16; i++)
-                    a[i] = (__bf16)smem_Q[qk_m + lane_lo][k_off + i];
+                    a[i] = (__bf16)(float)smem_Q[qk_m + lane_lo][k_off + i];
 
                 #pragma unroll
                 for (int nt = 0; nt < QK_N_TILES; nt++) {
@@ -248,7 +248,7 @@ extern "C" __global__ void KERNEL_NAME(
                     v16bf bb;
                     #pragma unroll
                     for (int k = 0; k < 16; k++)
-                        bb[k] = (__bf16)smem_K[key_row][k_off + k];
+                        bb[k] = (__bf16)(float)smem_K[key_row][k_off + k];
                     acc_s[nt] = __builtin_amdgcn_wmma_f32_16x16x16_bf16_w32(a, bb, acc_s[nt]);
                 }
             }
@@ -341,7 +341,7 @@ extern "C" __global__ void KERNEL_NAME(
                 v16bf a;
                 #pragma unroll
                 for (int i = 0; i < 16; i++)
-                    a[i] = (__bf16)smem_P[pv_warp_m + lane_lo][k_off + i];
+                    a[i] = (__bf16)(float)smem_P[pv_warp_m + lane_lo][k_off + i];
 
                 #pragma unroll
                 for (int nt = 0; nt < PV_N_TILES; nt++) {
@@ -349,7 +349,7 @@ extern "C" __global__ void KERNEL_NAME(
                     v16bf bb;
                     #pragma unroll
                     for (int k = 0; k < 16; k++)
-                        bb[k] = (__bf16)smem_K[k_off + k][d_col]; // buffer-share: smem_K holds V here
+                        bb[k] = (__bf16)(float)smem_K[k_off + k][d_col]; // buffer-share: smem_K holds V here
                     acc_o[nt] = __builtin_amdgcn_wmma_f32_16x16x16_bf16_w32(a, bb, acc_o[nt]);
                 }
 #else
@@ -529,14 +529,14 @@ extern "C" __global__ void PAGED_CONCAT(KERNEL_NAME, _64)(
                 v16bf a;
                 #pragma unroll
                 for (int i = 0; i < 16; i++)
-                    a[i] = (__bf16)smem_Q[qk_m + lane_lo][k_off + i];
+                    a[i] = (__bf16)(float)smem_Q[qk_m + lane_lo][k_off + i];
                 #pragma unroll
                 for (int nt = 0; nt < QK_N_TILES; nt++) {
                     unsigned int key_row = nt * 16 + lane_lo;
                     v16bf bb;
                     #pragma unroll
                     for (int k = 0; k < 16; k++)
-                        bb[k] = (__bf16)smem_K[key_row][k_off + k];
+                        bb[k] = (__bf16)(float)smem_K[key_row][k_off + k];
                     acc_s[nt] = __builtin_amdgcn_wmma_f32_16x16x16_bf16_w32(a, bb, acc_s[nt]);
                 }
             }
@@ -615,14 +615,14 @@ extern "C" __global__ void PAGED_CONCAT(KERNEL_NAME, _64)(
                 v16bf a;
                 #pragma unroll
                 for (int i = 0; i < 16; i++)
-                    a[i] = (__bf16)smem_P[pv_warp_m + lane_lo][k_off + i];
+                    a[i] = (__bf16)(float)smem_P[pv_warp_m + lane_lo][k_off + i];
                 #pragma unroll
                 for (int nt = 0; nt < PV_N_TILES; nt++) {
                     unsigned int d_col = (pv_n_start + nt) * 16 + lane_lo;
                     v16bf bb;
                     #pragma unroll
                     for (int k = 0; k < 16; k++)
-                        bb[k] = (__bf16)smem_K[k_off + k][d_col]; // buffer-share: smem_K holds V here
+                        bb[k] = (__bf16)(float)smem_K[k_off + k][d_col]; // buffer-share: smem_K holds V here
                     acc_o[nt] = __builtin_amdgcn_wmma_f32_16x16x16_bf16_w32(a, bb, acc_o[nt]);
                 }
 #else
