+++ kernels/strix-hip/qwen3.6-27b/nvfp4/w4a16_gemm.cu
@@ -159,12 +159,12 @@ extern "C" __global__ void w4a16_gemm(
 
         v16bf a;
         #pragma unroll
-        for (int i = 0; i < 16; i++) a[i] = (__bf16)smem_A[warp_m_offset + (lane_id & 15)][i];
+        for (int i = 0; i < 16; i++) a[i] = (__bf16)(float)smem_A[warp_m_offset + (lane_id & 15)][i];
         #pragma unroll
         for (int nb = 0; nb < 4; nb++) {
             v16bf b;
             #pragma unroll
-            for (int k = 0; k < 16; k++) b[k] = (__bf16)smem_B[k][nb * 16 + (lane_id & 15)];
+            for (int k = 0; k < 16; k++) b[k] = (__bf16)(float)smem_B[k][nb * 16 + (lane_id & 15)];
             acc[nb] = __builtin_amdgcn_wmma_f32_16x16x16_bf16_w32(a, b, acc[nb]);
         }
         __syncthreads();
@@ -270,14 +270,14 @@ extern "C" __global__ void w4a16_gemm_t(
             v16bf a; \
             _Pragma("unroll") \
             for (int i = 0; i < 16; i++) \
-                a[i] = (__bf16)smem_A[(a_buf)][warp_m_offset + (lane_id & 15)][h * 16 + i]; \
+                a[i] = (__bf16)(float)smem_A[(a_buf)][warp_m_offset + (lane_id & 15)][h * 16 + i]; \
             _Pragma("unroll") \
             for (int nb = 0; nb < 8; nb++) { \
                 unsigned int nc = nb * 16 + (lane_id & 15); \
                 v16bf b; \
                 _Pragma("unroll") \
                 for (int k = 0; k < 16; k++) \
-                    b[k] = (__bf16)smem_B_bf16[nc][h * 16 + k]; \
+                    b[k] = (__bf16)(float)smem_B_bf16[nc][h * 16 + k]; \
                 acc[nb] = __builtin_amdgcn_wmma_f32_16x16x16_bf16_w32(a, b, acc[nb]); \
             } \
         } \
@@ -371,14 +371,14 @@ extern "C" __global__ void fp8_gemm_t(
             v16bf a; \
             _Pragma("unroll") \
             for (int i = 0; i < 16; i++) \
-                a[i] = (__bf16)smem_A[(a_buf)][warp_m_offset + (lane_id & 15)][h * 16 + i]; \
+                a[i] = (__bf16)(float)smem_A[(a_buf)][warp_m_offset + (lane_id & 15)][h * 16 + i]; \
             _Pragma("unroll") \
             for (int nb = 0; nb < 8; nb++) { \
                 unsigned int nc = nb * 16 + (lane_id & 15); \
                 v16bf b; \
                 _Pragma("unroll") \
                 for (int k = 0; k < 16; k++) \
-                    b[k] = (__bf16)atlas_e4m3_to_f32(smem_B[(b_buf)][nc][h * 16 + k]); \
+                    b[k] = (__bf16)(float)atlas_e4m3_to_f32(smem_B[(b_buf)][nc][h * 16 + k]); \
                 acc[nb] = __builtin_amdgcn_wmma_f32_16x16x16_bf16_w32(a, b, acc[nb]); \
             } \
         } \
@@ -517,14 +517,14 @@ extern "C" __global__ void fp8_fp8_gemm_t(
             v16bf a; \
             _Pragma("unroll") \
             for (int i = 0; i < 16; i++) \
-                a[i] = (__bf16)atlas_e4m3_to_f32(smem_Af[(a_buf)][warp_m_offset + (lane_id & 15)][h * 16 + i]); \
+                a[i] = (__bf16)(float)atlas_e4m3_to_f32(smem_Af[(a_buf)][warp_m_offset + (lane_id & 15)][h * 16 + i]); \
             _Pragma("unroll") \
             for (int nb = 0; nb < 8; nb++) { \
                 unsigned int nc = nb * 16 + (lane_id & 15); \
                 v16bf b; \
                 _Pragma("unroll") \
                 for (int k = 0; k < 16; k++) \
-                    b[k] = (__bf16)atlas_e4m3_to_f32(smem_Bf[(b_buf)][nc][h * 16 + k]); \
+                    b[k] = (__bf16)(float)atlas_e4m3_to_f32(smem_Bf[(b_buf)][nc][h * 16 + k]); \
                 acc[nb] = __builtin_amdgcn_wmma_f32_16x16x16_bf16_w32(a, b, acc[nb]); \
             } \
         } \
@@ -668,14 +668,14 @@ extern "C" __global__ void w4a16_gemm_t_k64(
             v16bf a; \
             _Pragma("unroll") \
             for (int i = 0; i < 16; i++) \
-                a[i] = (__bf16)smem_A_k64[(a_buf)][warp_m_offset + (lane_id & 15)][h * 16 + i]; \
+                a[i] = (__bf16)(float)smem_A_k64[(a_buf)][warp_m_offset + (lane_id & 15)][h * 16 + i]; \
             _Pragma("unroll") \
             for (int nb = 0; nb < 8; nb++) { \
                 unsigned int nc = nb * 16 + (lane_id & 15); \
                 v16bf b; \
                 _Pragma("unroll") \
                 for (int k = 0; k < 16; k++) \
-                    b[k] = (__bf16)smem_B_bf16_k64[nc][h * 16 + k]; \
+                    b[k] = (__bf16)(float)smem_B_bf16_k64[nc][h * 16 + k]; \
                 acc[nb] = __builtin_amdgcn_wmma_f32_16x16x16_bf16_w32(a, b, acc[nb]); \
             } \
         } \
@@ -812,14 +812,14 @@ void w4a16_gemm_t_m128(
                 v16bf a; \
                 _Pragma("unroll") \
                 for (int i = 0; i < 16; i++) \
-                    a[i] = (__bf16)smem_A[(a_buf)][m_row][h * 16 + i]; \
+                    a[i] = (__bf16)(float)smem_A[(a_buf)][m_row][h * 16 + i]; \
                 _Pragma("unroll") \
                 for (int nb = 0; nb < 8; nb++) { \
                     unsigned int nc = nb * 16 + (lane_id & 15); \
                     v16bf b; \
                     _Pragma("unroll") \
                     for (int k = 0; k < 16; k++) \
-                        b[k] = (__bf16)smem_B_bf16[nc][h * 16 + k]; \
+                        b[k] = (__bf16)(float)smem_B_bf16[nc][h * 16 + k]; \
                     acc[nb] = __builtin_amdgcn_wmma_f32_16x16x16_bf16_w32(a, b, acc[nb]); \
                 } \
             } \
@@ -932,14 +932,14 @@ void fp8_gemm_t_m128(
                 v16bf a; \
                 _Pragma("unroll") \
                 for (int i = 0; i < 16; i++) \
-                    a[i] = (__bf16)smem_A[(a_buf)][m_row][h * 16 + i]; \
+                    a[i] = (__bf16)(float)smem_A[(a_buf)][m_row][h * 16 + i]; \
                 _Pragma("unroll") \
                 for (int nb = 0; nb < 8; nb++) { \
                     unsigned int nc = nb * 16 + (lane_id & 15); \
                     v16bf b; \
                     _Pragma("unroll") \
                     for (int k = 0; k < 16; k++) \
-                        b[k] = (__bf16)atlas_e4m3_to_f32(smem_B[(b_buf)][nc][h * 16 + k]); \
+                        b[k] = (__bf16)(float)atlas_e4m3_to_f32(smem_B[(b_buf)][nc][h * 16 + k]); \
                     acc[nb] = __builtin_amdgcn_wmma_f32_16x16x16_bf16_w32(a, b, acc[nb]); \
                 } \
             } \
@@ -1044,14 +1044,14 @@ void fp8_fp8_gemm_t_m128(
                 v16bf a; \
                 _Pragma("unroll") \
                 for (int i = 0; i < 16; i++) \
-                    a[i] = (__bf16)atlas_e4m3_to_f32(smem_Af[(a_buf)][m_row][h * 16 + i]); \
+                    a[i] = (__bf16)(float)atlas_e4m3_to_f32(smem_Af[(a_buf)][m_row][h * 16 + i]); \
                 _Pragma("unroll") \
                 for (int nb = 0; nb < 8; nb++) { \
                     unsigned int nc = nb * 16 + (lane_id & 15); \
                     v16bf b; \
                     _Pragma("unroll") \
                     for (int k = 0; k < 16; k++) \
-                        b[k] = (__bf16)atlas_e4m3_to_f32(smem_Bf[(b_buf)][nc][h * 16 + k]); \
+                        b[k] = (__bf16)(float)atlas_e4m3_to_f32(smem_Bf[(b_buf)][nc][h * 16 + k]); \
                     acc[nb] = __builtin_amdgcn_wmma_f32_16x16x16_bf16_w32(a, b, acc[nb]); \
                 } \
             } \
